+++ src/providers/image-loader-config.ts
@@ -1,56 +1,56 @@
-import { Injectable } from '@angular/core';
 import { HttpHeaders } from '@angular/common/http';
+import { Injectable }  from '@angular/core';
 
 @Injectable()
 export class ImageLoaderConfig {
+  debugMode = false;
 
-  debugMode: boolean = false;
+  spinnerEnabled = true;
 
-  spinnerEnabled: boolean = true;
+  fallbackAsPlaceholder = false;
 
-  fallbackAsPlaceholder: boolean = false;
+  backgroundSize = 'contain';
 
-  backgroundSize: string = 'contain';
+  backgroundRepeat = 'no-repeat';
 
-  backgroundRepeat: string = 'no-repeat';
+  display = 'block';
 
-  display: string = 'block';
+  width = '100%';
 
-  width: string = '100%';
+  height = '100%';
 
-  height: string = '100%';
+  useImg = false;
 
-  useImg: boolean = false;
+  fallbackUrl;
 
-  fallbackUrl: string;
+  concurrency = 5;
 
-  concurrency: number = 5;
+  maxCacheSize = -1;
 
-  maxCacheSize: number = -1;
-
-  maxCacheAge: number = -1;
+  maxCacheAge = -1;
 
   imageReturnType: 'base64' | 'uri' = 'uri';
 
-  spinnerName: string;
+  spinnerName;
 
-  spinnerColor: string;
+  spinnerColor;
 
   httpHeaders: HttpHeaders;
 
-  fileNameCachedWithExtension: boolean = false;
+  fileNameCachedWithExtension = false;
 
-  fallbackFileNameCachedExtension: string = '.jpg';
+  fallbackFileNameCachedExtension = '.jpg';
 
   cacheDirectoryType: 'cache'|'data' = 'cache';
 
   private _cacheDirectoryName: string = 'image-loader-cache';
 
-  get cacheDirectoryName(): string {
+
+  get cacheDirectoryName() {
     return this._cacheDirectoryName;
   }
 
-  set cacheDirectoryName(name: string) {
+  set cacheDirectoryName(name) {
     name.replace(/\W/g, '');
     this._cacheDirectoryName = name;
   }
@@ -64,7 +64,7 @@ export class ImageLoaderConfig {
 
   /**
    * Enable/Disable the spinner by default. Defaults to true.
-   * @param enable {boolean} set to true to enable
+   * @param {boolean} enable set to true to enable
    */
   enableSpinner(enable: boolean): void {
     this.spinnerEnabled = enable;
@@ -72,7 +72,7 @@ export class ImageLoaderConfig {
 
   /**
    * Enable/Disable the fallback image as placeholder instead of the spinner. Defaults to false.
-   * @param enable {boolean} set to true to enable
+   * @param {boolean} enable set to true to enable
    */
   enableFallbackAsPlaceholder(enable: boolean): void {
     this.fallbackAsPlaceholder = enable;
@@ -80,7 +80,7 @@ export class ImageLoaderConfig {
 
   /**
    * Sets the cache directory name. Defaults to 'image-loader-cache'
-   * @param name {string} name of directory
+   * @param {string} name name of directory
    */
   setCacheDirectoryName(name: string): void {
     this.cacheDirectoryName = name;
@@ -88,7 +88,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set default height for images that are not using <img> tag
-   * @param height {string} height
+   * @param {string} height height
    */
   setHeight(height: string): void {
     this.height = height;
@@ -96,7 +96,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set default width for images that are not using <img> tag
-   * @param width {string} Width
+   * @param {string} width Width
    */
   setWidth(width: string): void {
     this.width = width;
@@ -104,7 +104,7 @@ export class ImageLoaderConfig {
 
   /**
    * Enable display mode for images that are not using <img> tag
-   * @param display {string} Display mode
+   * @param {string} display Display mode
    */
   setDisplay(display: string): void {
     this.display = display;
@@ -112,7 +112,7 @@ export class ImageLoaderConfig {
 
   /**
    * Use <img> tag by default
-   * @param use {boolean} set to true to use <img> tag by default
+   * @param {boolean} use set to true to use <img> tag by default
    */
   useImageTag(use: boolean): void {
     this.useImg = use;
@@ -120,7 +120,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set default background size for images that are not using <img> tag
-   * @param backgroundSize {string} Background size
+   * @param {string} backgroundSize Background size
    */
   setBackgroundSize(backgroundSize: string): void {
     this.backgroundSize = backgroundSize;
@@ -128,7 +128,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set background repeat for images that are not using <img> tag
-   * @param backgroundRepeat {string} Background repeat
+   * @param {string} backgroundRepeat Background repeat
    */
   setBackgroundRepeat(backgroundRepeat: string): void {
     this.backgroundRepeat = backgroundRepeat;
@@ -137,7 +137,7 @@ export class ImageLoaderConfig {
   /**
    * Set fallback URL to use when image src is undefined or did not resolve.
    * This image will not be cached. This should ideally be a locally saved image.
-   * @param fallbackUrl {string} The remote or local URL of the image
+   * @param {string} fallbackUrl The remote or local URL of the image
    */
   setFallbackUrl(fallbackUrl: string): void {
     this.fallbackUrl = fallbackUrl;
@@ -145,7 +145,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set the maximum number of allowed connections at the same time.
-   * @param concurrency
+   * @param {number} concurrency
    */
   setConcurrency(concurrency: number): void {
     this.concurrency = concurrency;
@@ -153,7 +153,7 @@ export class ImageLoaderConfig {
 
   /**
    * Sets the maximum allowed cache size
-   * @param cacheSize {number} Cache size in bytes
+   * @param {number} cacheSize Cache size in bytes
    */
   setMaximumCacheSize(cacheSize: number): void {
     this.maxCacheSize = cacheSize;
@@ -161,7 +161,7 @@ export class ImageLoaderConfig {
 
   /**
    * Sets the maximum allowed cache age
-   * @param cacheAge {number} Maximum cache age in milliseconds
+   * @param {number} cacheAge Maximum cache age in milliseconds
    */
   setMaximumCacheAge(cacheAge: number): void {
     this.maxCacheAge = cacheAge;
@@ -176,8 +176,8 @@ export class ImageLoaderConfig {
   }
 
   /**
-   * Set the default spinnern ame
-   * @param name
+   * Set the default spinner name
+   * @param {string} name
    */
   setSpinnerName(name: string): void {
     this.spinnerName = name;
@@ -185,7 +185,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set the default spinner color
-   * @param color
+   * @param {string} color
    */
   setSpinnerColor(color: string): void {
     this.spinnerColor = color;
@@ -193,7 +193,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set headers options for the HttpClient transfers.
-   * @param headers
+   * @param {HttpHeaders} headers
    */
   setHttpHeaders(headers: HttpHeaders): void {
     this.httpHeaders = headers;
@@ -204,13 +204,16 @@ export class ImageLoaderConfig {
    * @param options
    * @deprecated FileTransfer plugin removed.
    */
-  setFileTransferOptions(options: { trustAllHosts: boolean; [key: string]: any; }): void {
+  setFileTransferOptions(options: {
+    trustAllHosts: boolean;
+    [key: string]: any;
+  }): void {
     // do nothing, plugin deprecated.
   }
 
   /**
    * Enable/Disable the save filename of cached images with extension.  Defaults to false.
-   * @param enable {boolean} set to true to enable
+   * @param {boolean} enable set to true to enable
    */
   setFileNameCachedWithExtension(enable: boolean) {
     this.fileNameCachedWithExtension = enable;
@@ -218,7 +221,7 @@ export class ImageLoaderConfig {
 
   /**
    * Set fallback extension filename of cached images.  Defaults to '.jpg'.
-   * @param extension {string} fallback extension (e.x .jpg)
+   * @param {string} extension fallback extension (e.x .jpg)
    */
   setFallbackFileNameCachedExtension(extension: string) {
     this.fallbackFileNameCachedExtension = extension;