+++ CHANGELOG.md
@@ -1,10 +1,51 @@
-<a name="5.0.5"></a>
-## [5.0.5](https://github.com/zyra/ionic-image-loader/compare/v5.0.4...v5.0.5) (2018-07-12)
+<a name="6.1.0"></a>
+# [6.1.0](https://github.com/zyra/ionic-image-loader/compare/v5.1.1...v6.1.0) (2018-09-17)
+
+
+### Features
+
+* **image-loader:** add ionic webView 2 support ([#189](https://github.com/zyra/ionic-image-loader/issues/189)) ([da035a8](https://github.com/zyra/ionic-image-loader/commit/da035a8)), closes [#187](https://github.com/zyra/ionic-image-loader/issues/187)
+* **lib:** Angular5 support ([cb8fd22](https://github.com/zyra/ionic-image-loader/commit/cb8fd22))
+
+
+
+<a name="5.1.1"></a>
+## [5.1.1](https://github.com/zyra/ionic-image-loader/compare/6.0.0...v5.1.1) (2018-09-17)
+
+
+### Bug Fixes
+
+* **image-loader:** fix rxjs fromEvent issue ([ce7f59f](https://github.com/zyra/ionic-image-loader/commit/ce7f59f))
+
+
+
+<a name="6.0.0"></a>
+# [6.0.0](https://github.com/zyra/ionic-image-loader/compare/v6.0.0...6.0.0) (2018-09-17)
+
+
+
+<a name="6.0.0"></a>
+# [6.0.0](https://github.com/zyra/ionic-image-loader/compare/5.1.0...v6.0.0) (2018-09-17)
+
+
+
+<a name="5.1.0"></a>
+# [5.1.0](https://github.com/zyra/ionic-image-loader/compare/v5.1.0...5.1.0) (2018-09-17)
+
+
+
+<a name="5.1.0"></a>
+# [5.1.0](https://github.com/zyra/ionic-image-loader/compare/v5.0.4...v5.1.0) (2018-09-17)
 
 
 ### Bug Fixes
 
 * **cache:** Images not saved to cache ([#171](https://github.com/zyra/ionic-image-loader/issues/171)) ([07eb2e7](https://github.com/zyra/ionic-image-loader/commit/07eb2e7))
+
+
+### Features
+
+* **img-loader:** pass attributes to img element ([#111](https://github.com/zyra/ionic-image-loader/issues/111)) ([482f6b5](https://github.com/zyra/ionic-image-loader/commit/482f6b5))